+++ api/src/routers/index.ts
@@ -1,0 +1,6 @@
+export { authenticationRouter } from './authentication';
+export { channelRouter } from './channel';
+export { channelInfoRouter } from './channel-info';
+export { identityRouter } from './identity';
+export { subscriptionRouter } from './subscription';
+export { verificationRouter } from './verification';
